feat: move film data ti index.tsx
