+++ project/src/components/SmallFilmCard/SmallFilmCard.types.ts
@@ -1,5 +1,4 @@
-export type Props = {
-  imgSrc: string;
-  name: string;
-  link: string;
-};
+/* eslint-disable quotes */
+import type { FilmItemType } from "../app/app.types";
+
+export type Props = Omit<FilmItemType, "id">;
