+++ project/src/components/app/app.types.ts
@@ -1,0 +1,10 @@
+export type FilmItemType = {
+  id: number;
+  imgSrc: string;
+  name: string;
+  link: string;
+};
+
+export type Props = {
+  films: FilmItemType[];
+};
